Remake application for sessionscope cart

+++ src/main/java/com/example/store/dto/CartAdditionDTO.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class CartAdditionDTO {
     private final Long id;
-    private final int quantity;
+    private final Long quantity;
 }
